@@ -1,55 +1,37 @@
 import { useState, useEffect } from "react";
-import MarvelService from "../../services/MarvelService";
 import Spinner from "../spinner/Spinner";
 import ErrorMessage from "../errorMessage/ErrorMessage";
+import useMarvelService from "../../services/MarvelService";
 
 import "./randomChar.scss";
 import mjolnir from "../../resources/img/mjolnir.png";
 
-const RandomChar = (props) => {
+const RandomChar = () => {
   const [char, setChar] = useState({});
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState(false);
-  const marvelService = new MarvelService();
+  const { loading, error, getCharacter, clearError } = useMarvelService();
 
-  // componentDidMount() {
-  // this.updateChar();
-  // this.timerID = setInterval(this.updateChar, 3000);
-  // }
-
-  // componentWillUnmount() {
-  //   clearInterval(this.timerID);
-  // }
   useEffect(() => {
     updateChar();
-    // const timerID = setInterval(updateChar, 50000);
-    // return () => clearInterval(timerID);
+    // const timerId = setInterval(updateChar, 60000);
+
+    // return () => {
+    //   clearInterval(timerId);
+    // };
   }, []);
 
-  const updateChar = () => {
-    const id = Math.floor(Math.random() * (1011400 - 1011000) + 1011000);
-    onCharLoading();
-    marvelService.getCharacter(id).then(onCharLoaded).catch(onError);
-  };
-
   const onCharLoaded = (char) => {
-    setLoading(false);
     setChar(char);
-    setError(false);
   };
 
-  const onCharLoading = () => {
-    setLoading(true);
-  };
-
-  const onError = () => {
-    setError(true);
-    setLoading(false);
+  const updateChar = () => {
+    clearError();
+    const id = Math.floor(Math.random() * (1011400 - 1011000)) + 1011000;
+    getCharacter(id).then(onCharLoaded);
   };
 
   const errorMessage = error ? <ErrorMessage /> : null;
   const spinner = loading ? <Spinner /> : null;
-  const content = !(loading || error) ? <View char={char} /> : null;
+  const content = !(loading || error || !char) ? <View char={char} /> : null;
 
   return (
     <div className="randomchar">
@@ -63,7 +45,7 @@ const RandomChar = (props) => {
           Do you want to get to know him better?
         </p>
         <p className="randomchar__title">Or choose another one</p>
-        <button className="button button__main" onClick={updateChar}>
+        <button onClick={updateChar} className="button button__main">
           <div className="inner">try it</div>
         </button>
         <img src={mjolnir} alt="mjolnir" className="randomchar__decoration" />
@@ -79,7 +61,7 @@ const View = ({ char }) => {
     thumbnail ===
     "http://i.annihil.us/u/prod/marvel/i/mg/b/40/image_not_available.jpg"
   ) {
-    imgStyle = { objectFit: "unset" };
+    imgStyle = { objectFit: "contain" };
   }
 
   return (
